나눗셈 함수 double형으로 변환

+++ Homework.c
@@ -15,7 +15,7 @@ int Mul(int num1 , int num2)
  return num1 * num2;
 }
 
-int division(int num1, int num2){
+double division(double num1, double num2){
  return num1 / num2;
 }
